examples/durable/durable_hello_sequence/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "durable_hello_sequence"
version = "0.1.0"
description = "Durable Functions orchestration chaining activity calls in sequence."
requires-python = ">=3.10"
dependencies = [
  "azure-functions>=1.21.0",
  "azure-functions-durable>=1.2.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
